Models adjustments
price double => decimal, removed join table, connected db

+++ Wise_Owl_Library/Data/Dto/Requests/CreateBookDto.cs
@@ -8,8 +8,8 @@ namespace Wise_Owl_Library.Data.Dto.Requests
         [StringLength(100, ErrorMessage = "Title length can't be more than 100 characters.")]
         public required string Title { get; set; }
 
-        [Range(0, double.MaxValue, ErrorMessage = "Price must be a positive value.")]
-        public double Price { get; set; }
+        [Range(0.0, double.MaxValue, ErrorMessage = "Price must be a positive value.")]
+        public decimal Price { get; set; }
 
         [Range(0, 10000, ErrorMessage = "Stock must be between 0 and 10000.")]
         public int Stock { get; set; }
